+++ TaggerV1/interfaces/iapifactory.ts
@@ -1,0 +1,7 @@
+import { IBuildApiRetry } from "./ibuildapiretry";
+
+export interface IApiFactory {
+
+    createBuildApi(): Promise<IBuildApiRetry>;
+
+}