+++ TaggerV1/interfaces/iparameters.ts
@@ -1,0 +1,9 @@
+import { IArtifact } from "./iartifact";
+
+export interface IParameters {
+
+    artifacts: IArtifact[];
+    tags: string[];
+    remove: boolean;
+
+}