+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.0
+
+- Add multiple artifacts support ([#6](https://github.com/dmitryserbin/azdev-artifacts-tagger/issues/6))
+- Add external artifact project support
+- Rafactor internal modules structure
+- Refactor retry mechanism
+- Update task to Node 10
+
 ## 1.1.2
 
 - Address endpoint not present issue ([#3](https://github.com/dmitryserbin/azdev-artifacts-tagger/issues/3))